+++ src/Humanizer/Configuration/LocaliserRegistry.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
 
 namespace Humanizer.Configuration
@@ -7,16 +8,26 @@ namespace Humanizer.Configuration
     /// A registry of localised system components with their associated locales
     /// </summary>
     /// <typeparam name="TLocaliser"></typeparam>
-    public class LocaliserRegistry<TLocaliser> 
+    public class LocaliserRegistry<TLocaliser>
+        where TLocaliser : class
     {
-        private readonly IDictionary<string, TLocaliser> _localisers = new Dictionary<string, TLocaliser>();
-        private readonly TLocaliser _defaultLocaliser;
+        private readonly IDictionary<string, Func<CultureInfo, TLocaliser>> _localisers = new Dictionary<string, Func<CultureInfo, TLocaliser>>();
+        private readonly Func<CultureInfo, TLocaliser> _defaultLocaliser;
 
         /// <summary>
         /// Creates a localiser registry with the default localiser set to the provided value
         /// </summary>
         /// <param name="defaultLocaliser"></param>
         public LocaliserRegistry(TLocaliser defaultLocaliser)
+        {
+            _defaultLocaliser = (culture) => defaultLocaliser;
+        }
+
+        /// <summary>
+        /// Creates a localiser registry with the default localiser factory set to the provided value
+        /// </summary>
+        /// <param name="defaultLocaliser"></param>
+        public LocaliserRegistry(Func<CultureInfo, TLocaliser> defaultLocaliser)
         {
             _defaultLocaliser = defaultLocaliser;
         }
@@ -35,9 +46,28 @@ namespace Humanizer.Configuration
         /// <param name="culture">The culture to retrieve localiser for. If not specified, current thread's UI culture is used.</param>
         public TLocaliser ResolveForCulture(CultureInfo culture)
         {
-            culture = culture ?? CultureInfo.CurrentUICulture;
+            return FindLocaliser(culture ?? CultureInfo.CurrentUICulture)(culture);
+        }
 
-            TLocaliser localiser;
+        /// <summary>
+        /// Registers the localiser for the culture provided 
+        /// </summary>
+        public void Register(string localeCode, TLocaliser localiser)
+        {
+            _localisers[localeCode] = (culture) => localiser;
+        }
+
+        /// <summary>
+        /// Registers the localiser factory for the culture provided
+        /// </summary>
+        public void Register(string localeCode, Func<CultureInfo, TLocaliser> localiser)
+        {
+            _localisers[localeCode] = localiser;
+        }
+
+        private Func<CultureInfo, TLocaliser> FindLocaliser(CultureInfo culture)
+        {
+            Func<CultureInfo, TLocaliser> localiser;
 
             if (_localisers.TryGetValue(culture.Name, out localiser))
                 return localiser;
@@ -46,14 +76,6 @@ namespace Humanizer.Configuration
                 return localiser;
 
             return _defaultLocaliser;
-        }
-
-        /// <summary>
-        /// Registers the localiser for the culture provided 
-        /// </summary>
-        public void Register(string localeCode, TLocaliser localiser)
-        {
-            _localisers[localeCode] = localiser;
         }
     }
 }