+++ src/Humanizer/Configuration/NumberToWordsConverterRegistry.cs
@@ -4,19 +4,19 @@ namespace Humanizer.Configuration
 {
     internal class NumberToWordsConverterRegistry : LocaliserRegistry<INumberToWordsConverter>
     {
-        public NumberToWordsConverterRegistry() : base(new DefaultNumberToWordsConverter())
+        public NumberToWordsConverterRegistry() : base((culture) => new DefaultNumberToWordsConverter(culture))
         {
             Register("en", new EnglishNumberToWordsConverter());
             Register("ar", new ArabicNumberToWordsConverter());
             Register("fa", new FarsiNumberToWordsConverter());
             Register("es", new SpanishNumberToWordsConverter());
-            Register("pl", new PolishNumberToWordsConverter());
+            Register("pl", (culture) => new PolishNumberToWordsConverter(culture));
             Register("pt-BR", new BrazilianPortugueseNumberToWordsConverter());
             Register("ru", new RussianNumberToWordsConverter());
             Register("fr", new FrenchNumberToWordsConverter());
             Register("nl", new DutchNumberToWordsConverter());
-            Register("he", new HebrewNumberToWordsConverter());
-            Register("sl", new SlovenianNumberToWordsConverter());
+            Register("he", (culture) => new HebrewNumberToWordsConverter(culture));
+            Register("sl", (culture) => new SlovenianNumberToWordsConverter(culture));
             Register("de", new GermanNumberToWordsConverter());
         }
     }