+++ src/Humanizer/Localisation/NumberToWords/GenderlessNumberToWordsConverter.cs
@@ -1,0 +1,41 @@
+﻿namespace Humanizer.Localisation.NumberToWords
+{
+    abstract class GenderlessNumberToWordsConverter : INumberToWordsConverter
+    {
+        /// <summary>
+        /// Converts the number to string
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns></returns>
+        public abstract string Convert(int number);
+
+        /// <summary>
+        /// Converts the number to string ignoring the provided grammatical gender
+        /// </summary>
+        /// <param name="number"></param>
+        /// <param name="gender"></param>
+        /// <returns></returns>
+        public string Convert(int number, GrammaticalGender gender)
+        {
+            return Convert(number);
+        }
+
+        /// <summary>
+        /// Converts the number to ordinal string
+        /// </summary>
+        /// <param name="number"></param>
+        /// <returns></returns>
+        public abstract string ConvertToOrdinal(int number);
+
+        /// <summary>
+        /// Converts the number to ordinal string ignoring  the provided grammatical gender
+        /// </summary>
+        /// <param name="number"></param>
+        /// <param name="gender"></param>
+        /// <returns></returns>
+        public string ConvertToOrdinal(int number, GrammaticalGender gender)
+        {
+            return ConvertToOrdinal(number);
+        }
+    }
+}